// backend/src/db/schemas/credential-logins.ts
// Code generated by automation script, DO NOT EDIT.
// Automated by pulling database and generating zod schema
// To update. Just run npm run generate:schema
// Written by akhilmhdh.

import { z } from "zod";

import { TImmutableDBKeys } from "./models";

export const CredentialLoginsSchema = z.object({
  id: z.string().uuid(),
  createdAt: z.date(),
  updatedAt: z.date(),
  orgId: z.string().uuid(),
  userId: z.string().uuid(),
  name: z.string(),
  website: z.string().nullable().optional(),
  username: z.string(),
  password: z.string()
});

export type TCredentialLogins = z.infer<typeof CredentialLoginsSchema>;
export type TCredentialLoginsInsert = Omit<z.input<typeof CredentialLoginsSchema>, TImmutableDBKeys>;
export type TCredentialLoginsUpdate = Partial<Omit<z.input<typeof CredentialLoginsSchema>, TImmutableDBKeys>>;